requirements.txt
uvicorn[standard]
gunicorn
fastapi[all]
celery[redis]
pydantic-settings
cartopy
owslib
